requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   starlette
    #   watchfiles
asyncpg==0.30.0
    # via demo-project1 (pyproject.toml)
click==8.1.8
    # via uvicorn
fastapi==0.115.12
    # via demo-project1 (pyproject.toml)
h11==0.14.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via anyio
pydantic==2.10.6
    # via
    #   demo-project1 (pyproject.toml)
    #   fastapi
pydantic-core==2.27.2
    # via pydantic
python-dotenv==1.1.0
    # via uvicorn
pyyaml==6.0.2
    # via uvicorn
sniffio==1.3.1
    # via anyio
starlette==0.46.1
    # via fastapi
typing-extensions==4.13.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn[standard]==0.34.0
    # via demo-project1 (pyproject.toml)
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.4
    # via uvicorn
websockets==15.0.1
    # via uvicorn
